♻️  refact(firebase-database): Changed the initialziation of the app instance
There was an internal error  when invoking functions. The bug was triggered by trying to initialize the app (Firebase-admin sdk) in each function. To tackle the problem a new file was created firebaseConfig.js that handles a unique initialization of the admin sdk that will be used in all of the fuinctions